+++ src/css/css-url.js
@@ -1,6 +1,7 @@
 import sass from 'sass';
 
 import AbstractCssProperty from './abstract-css-property';
+import QueryConstant from '../query-constant';
 
 export default class CssUrl extends AbstractCssProperty {
   /**
@@ -8,38 +9,119 @@ export default class CssUrl extends AbstractCssProperty {
    * @private
    */
   _url = undefined;
+  /**
+   * @type {boolean}
+   * @private
+   */
+  _inline = undefined;
 
   /**
    * @param {string} url
+   * @param {boolean} [inline=false]
    */
-  constructor(url) {
+  constructor(url, inline) {
     super();
     /**
      * @type {string}
      * @private
      */
     this._url = url;
+    /**
+     * @type {boolean}
+     * @private
+     */
+    this._inline = !!inline;
   }
 
   /**
    * @return {string}
    */
-  get url() {
+  get rawUrl() {
     return this._url;
   }
 
   /**
    * @return {string}
    */
+  get externalUrl() {
+    return this.rawUrl;
+  }
+
+  /**
+   * @return {string}
+   */
+  get inlineUrl() {
+    return this._getInlineUrl();
+  }
+
+  /**
+   * @return {string}
+   */
+  get url() {
+    return this.inline ? this.inlineUrl : this.rawUrl;
+  }
+
+  /**
+   * @return {string}
+   */
   get ref() {
-    return `@ref(${this.url})`;
+    return this._getRef(this.url);
+  }
+
+  /**
+   * @return {string}
+   */
+  get inlineRef() {
+    return this._getRef(this.inlineUrl);
+  }
+
+  /**
+   * @return {string}
+   */
+  get externalRef() {
+    return this._getRef(this.rawUrl);
   }
 
   /**
    * @return {string}
    */
   get css() {
-    return `url(${this.url})`;
+    return this._getCss(this.url);
+  }
+
+  /**
+   * @return {string}
+   */
+  get inlineCss() {
+    return this._getCss(this.inlineUrl);
+  }
+
+  /**
+   * @return {string}
+   */
+  get inlineRefCss() {
+    return this._getCss(this.inlineRef);
+  }
+
+  /**
+   * @return {string}
+   */
+  get externalCss() {
+    return this._getCss(this.rawUrl);
+  }
+
+  /**
+   * @return {string}
+   */
+  get externalRefCss() {
+    return this._getCss(this.externalRef);
+  }
+
+  /**
+   * @return {boolean}
+   */
+  get inline() {
+    return this._inline;
   }
 
   /**
@@ -61,6 +143,32 @@ export default class CssUrl extends AbstractCssProperty {
    */
   toString() {
     return `url(${this.ref})`;
+  }
+
+  /**
+   * @return {string}
+   * @private
+   */
+  _getInlineUrl() {
+    return `${this.rawUrl}?${QueryConstant.INLINE}`; // FIXME: care about already applied query strings
+  }
+
+  /**
+   * @param {string} url
+   * @return {string}
+   * @private
+   */
+  _getRef(url) {
+    return `@ref(${url})`;
+  }
+
+  /**
+   * @param {string} url
+   * @return {string}
+   * @private
+   */
+  _getCss(url) {
+    return `url(${url})`;
   }
 
   /**