+++ src/query-constant.js
@@ -1,0 +1,6 @@
+export default class QueryConstant {
+  /**
+   * @type {string}
+   */
+  static INLINE = 'inline';
+}